First version of custom ensemble

+++ src/extraction_benchmark/util.py
@@ -18,6 +18,8 @@ import requests
 from bs4 import BeautifulSoup
 import os
 import hashlib
+import csv
+import shutil
 
 def read_jsonl(file):
     """
@@ -80,35 +82,77 @@ def extract_urls_from_file(file):
     :param file: input file name
     :return: list of urls
     """
-
-    #TODO: add file detection option
     urls = []
-    with open(file) as f:
-        for line in f:
-            urls.append(line.strip())
+    #TODO: add file detection option
+    if file[-4:] == '.txt':
+        with open(file) as f:
+            for line in f:
+                urls.append(line.strip())
+
+    elif file[-4:] == '.csv':
+        with open(file, newline='') as f:
+            reader = csv.reader(f)
+            for row in reader:
+                urls.append(row[0])
+    
+    elif file[-6:] == '.jsonl':
+        with open(file, 'r') as f:
+            for line in f:
+                data = json.loads(line)
+                url = data.get('url')  
+                if url:
+                    urls.append(url)
+
     return urls
 
+def get_hash_from_file(file:str):
+    #Hash the file content to get file id
+    m = hashlib.sha256()
+    with open(file, 'rb') as f:
+        m.update(f.read())
+    file_id = m.hexdigest()
+    return file_id
+
+def process_html_files(in_dir:str, out_dir:str):
+    for html in os.listdir(in_dir):
+        file_id = get_hash_from_file(os.path.join(in_dir, html))
+        print('file_id is ',file_id)
+        shutil.copy2(os.path.join(in_dir, html), os.path.join(out_dir, file_id+'.html'))
+
+def get_hash_from_url(url:str):
+    #Hash the file content to get file id
+    m = hashlib.sha256()
+    #m.update(response.content)
+    m.update(url.encode('utf-8'))
+    file_id = m.hexdigest()
+    return file_id
 
 def extract_html_from_urls(urls:list, path:str):
     custom_path = os.path.join(path, 'custom')
     association_path = os.path.join(path, 'url_association')
+    rvalue = True
     for url in urls:
+        try:
+            response = requests.get(url)
+            #response.raise_for_status()  # This will raise an error if the request was unsuccessful
 
-        response = requests.get(url)
-        response.raise_for_status()  # This will raise an error if the request was unsuccessful
 
-        #Hash the file content to get file id
-        m = hashlib.sha256()
-        m.update(response.content)
-        file_id = m.hexdigest()
-        file = os.path.join(custom_path, file_id + '.html')
-        association = os.path.join(association_path, file_id + '.txt')
+            file = os.path.join(custom_path, file_id + '.html')
 
-        soup = BeautifulSoup(response.content, 'html.parser')
+            if os.path.isfile(file):
+                print('File already exists')
+                continue
+            
+            file_id = get_hash_from_url(url)
+            association = os.path.join(association_path, file_id + '.txt')
 
-        with open(file, 'w', encoding='utf-8') as file:
-            file.write(soup.prettify())
+            soup = BeautifulSoup(response.content, 'html.parser')
 
-        with open(association, 'w', encoding='utf-8') as file:
-            file.write(url)
+            with open(file, 'w', encoding='utf-8') as file:
+                file.write(soup.prettify())
 
+            with open(association, 'w', encoding='utf-8') as file:
+                file.write(url)
+        except:
+            print('Failed to extract url:', url)
+    return rvalue
